Adding Routez pages with Controller - Adding validations forms

+++ routes/web.php
@@ -5,6 +5,7 @@ use App\Http\Controllers\CobrosController;
 use App\Http\Controllers\BeneficiariosController;
 use App\Http\Controllers\PdfController;
 use App\Http\Controllers\ProfileController;
+use App\Http\Controllers\ReportesController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/', function () {
@@ -21,26 +22,34 @@ Route::middleware('auth')->group(function () {
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 });
 
-Route::get('/cobros', [CobrosController::class, 'index'])->name('cobros.index');
+Route::get('/cobros', [CobrosController::class, 'index'])->middleware(['auth', 'verified', 'can:cobros.index'])->name('cobros.index');
 
-Route::get('/cobros/{id}', [CobrosController::class, 'show'])->name('cobros.show');
+Route::get('/cobros/{id}', [CobrosController::class, 'show'])->middleware(['auth', 'verified', 'can:cobros.show'])->name('cobros.show');
 
-Route::post('/cobros/{beneficiario}', [CobrosController:: class, 'pagar'])->name('cobros.pagar');
+Route::post('/cobros/{beneficiario}', [CobrosController:: class, 'pagar'])->middleware(['auth', 'verified', 'can:cobros.pagar'])->name('cobros.pagar');
 
-Route::get('/cobros/tiket/{beneficiario}', [PdfController::class, 'ticket'])->name('cobros.ticket');
+Route::get('/cobros/{beneficiario}/tarjeta', [CobrosController:: class, 'imprimir'])->middleware(['auth', 'verified', 'can:cobros.imprimir'])->name('cobros.imprimir');
 
-Route::get('/beneficiarios', [BeneficiariosController::class, 'index'])->name('beneficiarios.index');
+Route::get('/beneficiarios', [BeneficiariosController::class, 'index'])->middleware(['auth', 'verified', 'can:beneficiarios.index'])->name('beneficiarios.index');
 
-Route::get('/beneficiarios/registro', [BeneficiariosController::class, 'create'])->name('beneficiarios.register');
+Route::get('/beneficiarios/registro', [BeneficiariosController::class, 'create'])->middleware(['auth', 'verified', 'can:beneficiarios.register'])->name('beneficiarios.register');
 
-Route::post('/beneficiarios', [BeneficiariosController::class, 'store'])->name('beneficiarios.create');
+Route::post('/beneficiarios', [BeneficiariosController::class, 'store'])->middleware(['auth', 'verified', 'can:beneficiarios.create'])->name('beneficiarios.create');
 
-Route::get('/beneficiarios/{beneficiario}/edit', [BeneficiariosController::class, 'edit'])->name('beneficiarios.edit');
+Route::get('/beneficiarios/{beneficiario}/edit', [BeneficiariosController::class, 'edit'])->middleware(['auth', 'verified', 'can:beneficiarios.edit'])->name('beneficiarios.edit');
 
-Route::put('/beneficiarios/{beneficiario}', [BeneficiariosController::class, 'update'])->name('beneficiarios.update');
+Route::put('/beneficiarios/{beneficiario}', [BeneficiariosController::class, 'update'])->middleware(['auth', 'verified', 'can:beneficiarios.update'])->name('beneficiarios.update');
 
-Route::delete('/beneficiarios/{beneficiario}', [BeneficiariosController::class, 'destroy'])->name('beneficiarios.destroy');
+Route::delete('/beneficiarios/{beneficiario}', [BeneficiariosController::class, 'destroy'])->middleware(['auth', 'verified', 'can:beneficiarios.destroy'])->name('beneficiarios.destroy');
 
-Route::resource('users', UserController::class)->names('admin.users');
+Route::resource('users', UserController::class)->middleware(['auth', 'verified', 'can:admin.users.index'])->names('admin.users');
+
+Route::get('/reportes', [ReportesController::class, 'index'])->middleware(['auth', 'verified', 'can:reportes.index'])->name('reportes.index');
+
+Route::post('/reportes/dia', [ReportesController::class, 'dia'])->middleware(['auth', 'verified', 'can:reportes.dia'])->name('reportes.dia');
+
+Route::post('/reportes/mes', [ReportesController::class, 'mes'])->middleware(['auth', 'verified', 'can:reportes.mes'])->name('reportes.mes');
+
+Route::get('/reportes/users', [ReportesController::class, 'users'])->middleware(['auth', 'verified', 'can:reportes.users'])->name('reportes.users');
 
 require __DIR__.'/auth.php';
